memoize relation tree fetching
