+++ uiCommon/explorer/index.tsx
@@ -1,41 +1,18 @@
 import React, { useState } from "react";
 import { AbstractInterpreter } from "../../core/abstractInterpreter";
-import { Term, UserError } from "../../core/types";
+import { Term } from "../../core/types";
 import { noHighlight, HighlightProps } from "../dl/term";
 import { useJSONLocalStorage } from "../generic/hooks";
 import { RelationTree } from "./relationTree";
 import { VizArea } from "./vizArea";
-import { ensurePresent, sortBy } from "../../util/util";
+import { ensurePresent } from "../../util/util";
 import { OpenRelationsContainer } from "./openRelationsContainer";
-import { RelationCollapseStates, RelationInfo, RelationStatus } from "./types";
+import { RelationCollapseStates } from "./types";
 
 export function Explorer(props: {
   interp: AbstractInterpreter;
   showViz?: boolean;
 }) {
-  const allRules: RelationInfo[] = sortBy(
-    props.interp.getRules(),
-    (r) => r.head.relation
-  ).map((rule) => ({
-    type: "Rule",
-    name: rule.head.relation,
-    rule,
-    status: getStatus(props.interp, rule.head.relation),
-  }));
-  const allTables: RelationInfo[] = [
-    ...props.interp
-      .getTables()
-      .sort()
-      .map(
-        (name): RelationInfo => ({
-          type: "Table",
-          name,
-          status: getStatus(props.interp, name),
-        })
-      ),
-    // TODO: virtual tables
-  ];
-
   const [highlight, setHighlight] = useState(noHighlight);
 
   const [relationCollapseStates, setRelationCollapseStates] =
@@ -70,8 +47,6 @@ export function Explorer(props: {
       >
         <RelationTree
           interp={props.interp}
-          allRules={allRules}
-          allTables={allTables}
           highlight={highlightProps}
           openRelations={openRelations}
           setOpenRelations={setOpenRelations}
@@ -101,18 +76,4 @@ export function Explorer(props: {
       ) : null}
     </div>
   );
-}
-
-function getStatus(interp: AbstractInterpreter, name: string): RelationStatus {
-  try {
-    const count = interp.queryStr(`${name}{}`).length;
-    return { type: "Count", count };
-  } catch (e) {
-    if (!(e instanceof UserError)) {
-      console.error(`error while getting ${name}:`, e);
-    }
-    // TODO: this could swallow up an internal error...
-    // should get better about errors...
-    return { type: "Error" };
-  }
 }
